Fixed aproximation with strcmp()
Because the n'th position was afterwards compared between both strings, a str-
ing such as "ac" could pass evaluation with a string such as "bc". This is due
to disregard of a case where the  first character of the first string is lexi-
graphically less than the first character of the second string.